move delegate attribute tests
